Add problem 12 solutions

+++ project.euler/src/test/java/edo/project/euler/problem/Problem12Test.java
@@ -10,7 +10,7 @@ public class Problem12Test extends TestCase {
 	@Test
 	public void testBruteForce() {
 		int expected = -1;
-		int actual = new Problem12().solve();
+		long actual = new Problem12().solve();
 		assertEquals(expected, actual);
 	}
